refactor: 클러스터링 로직 커스텀훅으로 분리 (#174)
